+++ airbox/commands/run_schedule.py
@@ -29,10 +29,6 @@ class RunScheduleCommand(BaseCommand):
     name = 'run_schedule'
     help = 'Run the scheduler. This runs the tasks that need to be run daily or every hour'
 
-    def initialise_parser(self, subparser):
-        subparser.add_argument('--log-file', help="Filename where the logs from the scheduler are written. This file is"
-                                                  " appended to every invocation", default="/var/log/airbox.log")
-
     def run(self):
         # Late importing to resolve circular dependency
         from airbox.commands import run_command
@@ -55,5 +51,5 @@ class RunScheduleCommand(BaseCommand):
                 exc_info = sys.exc_info()
                 logger.exception('Command failed. Emailing expeds')
                 sendmail(config['email_expeditioners'], "Scheduled command failed: {}".format(s['command']),
-                         "{}. Check the airbox log (typically /var/log/airbox.log) on the airbox server for more "
+                         "{}. Check the airbox log (typically in /var/log/airbox/) on the airbox server for more "
                          "information.".format(exc_info[1]))